requirements.txt
openai
deta
streamlit_option_menu==0.3.6